rewrite code from Hibernate class to ORMManager
